Add files via upload
In this program, you can create different food packages with names and tastes. When a user selects a package to consume, it checks if consuming that package along with any previously consumed packages would result in a mixture of different tastes that could cause nausea. If there's a risk of nausea, it prompts the user to choose another package.

You can expand the list of food packages by adding more items to the create_food_packages function. This program helps users manage their food intake to avoid unpleasant taste combinations.